Public API

Public Documentation

Documentation for LTspice.jl's public interface

Index

LTspice

LTspice.LTspiceSimulation

Base.flush

LTspice.circuitpath

LTspice.executablepath

LTspice.logpath

LTspice.measurementnames

LTspice.measurementvalues

LTspice.parameternames

LTspice.parametervalues

LTspice.parselog!

LTspice.perlineiterator

LTspice.run!

LTspice.stepnames

LTspice.stepvalues

Working With Simulations

LTspice — Module.

Main module for LTspice.jl - a Julia interface to LTspice

source

LTspice.LTspiceSimulation — Type.

Access parameters and measurements of an LTspice simulation. Runs simulation as needed.

Access as a dictionary:

measurement_value = sim["measurement_name"]
parameter_value = sim["parameter_name"]
sim["parameter_name"] = new_parameter_value

Access as a function:

(m1,m2,m3) = sim(p1,p2,p3)  # simulation with three measurements and three parameters

Access as arrays or tuples:

pnames = parameternames(sim)
mnames = measurementnames(sim)
snames = stepnames(sim)
pvalues = parametervalues(sim)
mvalues = measurementvalues(sim)
svalues = stepvalues(sim)

source

LTspice.parametervalues — Function.

parametervalues(sim)

Returns an array of the parameters of sim in the order they appear in the circuit file

source

LTspice.parameternames — Function.

parameternames(sim)

Returns an tuple of the parameters names of sim in the order they appear in the circuit file.

source

LTspice.measurementvalues — Function.

measurementvalues(sim)

Retruns measurements of sim as an a array of Float64 values.

value = measurementvalues(sim)[inner_step,
                               middle_step,
                               outer_step,
                               measurement_name] # 3 nested steps

source

LTspice.measurementnames — Function.

measurementnames(sim)

Returns an tuple of the measurement names of sim in the order they appear in the circuit file.

source

LTspice.stepvalues — Function.

stepvalues(sim)

Returns the steps of sim as a tuple of three arrays of the step values.

source

LTspice.stepnames — Function.

stepnames(sim)

Returns an tuple of step names of sim.

source

LTspice.circuitpath — Function.

circuitpath(sim)

Returns path to the circuit file.

This is the path to the working circuit file. If tempdir=ture was used or if running under wine, this will not be the path given to the constructor.

source

LTspice.executablepath — Function.

executablepath(sim)

Returns path to the LTspice executable

source

LTspice.logpath — Function.

logpath(sim)

Returns path to the log file.

source

LTspice.parselog! — Function.

parselog!(sim)

Loads log file of sim without running simulation. The user does not normally need to call parselog!.

source

Base.flush — Function.

flush(sim)

Writes sim's circuit file back to disk if any parameters have changed. The user does not usually need to call flush. It will be called automatically when a measurement is requested and the log file needs to be updated. It can be used to update a circuit file using julia for simulation with the LTspice GUI.

source

LTspice.run! — Function.

run!(sim)

Writes circuit changes, calls LTspice to run sim, and reloads the log file. The user normally does not need to call this.

source

Exporting Data

LTspice.perlineiterator — Function.

perlineiterator(simulation, <keyword arguments>)

Retruns iterator that flattens multidimensional data.

Keyword Arguments

steporder – specify order of steps

resultnames – specify parameters and measurements for output

header – true to make first line header

The step order defaults to the order the step values appear in the circuit file. Step order can be specified by passing an array of step names. By default there is one column for each step, measurement, and parameter. The desired measurements and parameters can be set by passing an array of names to resultnames.

# write CSV with headers
using DelimitedFiles
writedlm("test.csv",perlineiterator(circuit2,header=true),',')

source